+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [0.3.0-beta.2] - 2023-06-21
+
+* Fix handling stream capacity
+
+* use ContainerCall instead of ServiceCall
+
 ## [0.3.0-beta.1] - 2023-06-19
 
 * Use ServiceCtx instead of Ctx